Remove unnecessary optimisation
There are two optimisations in the original solution,
and only one ends up being necessary.

Optimisation A: De-duplicating crabs at the same starting
position, then multiplying the cost by the number of crabs.

Optimisation B: Using Gauss' method to calculate the sum of
the sequence in part 2 instead of iterating through a
sequence and using sum().

Removing optimisation A has no noticable impact on
performance, with or without optimisation B.

Removing optimisation B does have a noticable impact on
performance for part 2, with or without optimisation A.

Therefore, remove optimisation A.

This makes the code shorter and less confusing. as it avoids
a Map<Int,Int> where each Int has a very different meaning —
starting position vs number of crabs.