Day 8, part 2
The logic to decode the outputs is tricky, but by finding options
of the right size that "don't contain" 1 or 7 in them (and some
"clever" work comparing all of the 5 and 6 character inputs to
each other, a stable algorithm exists.

The fact that find() just returns the first match if there's
more than one concealed a bug at first. Filter + single is more
verbose but better because it fails fast.