Day 14, part 2
Cheat outrageously (although I'm sure that's the idea) and instead
of calculating what the output polymer actually is, just calculate
how many of each kind of element are in it.

Keep the old PolymerTemplate around as an oracle to compare
results for the first 10 rounds, although strictly speaking we
could replace it and rewrite a bunch of tests.