Day 22, Part 1
Builf a cuboid difference calculator to split overlapping cuboids
into a list of cuboids that *don't* overlap, and then repeatedly
apply this to the list of cuboids, throwing away the "off"
operations after they have knocked holes into the cuboids from
the "on" operations. The result is a list of "on" cuboids that
do not intersect, so the answer is just the sum of the size of
those cuboids.

Figuring out a correct algorithm for calculating the difference
between two cuboids was a *very* painful, but once that is
in place the algorithm generates fast answers.